refactor: return component props to component files
