+++ packages/plugins/text/src/types/math-element.ts
@@ -1,11 +1,3 @@
-import { RenderElementProps } from 'slate-react'
-
-export interface MathElementProps {
-  element: MathElement
-  attributes: RenderElementProps['attributes']
-  children: RenderElementProps['children']
-}
-
 // TODO: Good structure
 export interface MathElement {
   type: 'math'
